svd each factor (jed)

+++ scripts/redownload_error_chip_data.sh
@@ -1,34 +1,43 @@
 #!/bin/bash
 
+genome=$1
+infile=$2
 
-Genome="hg38"
-
-infile="resources/to_dowload_${Genome}.txt"
 mapfile -t to_download  < <(cat "$infile")
 
-# get files
-for Experimental_ID in "${to_download[@]}"
-do
+# pass if file is empty
+if [ ${#to_download[@]} -eq 0 ]
+then
+    echo "No files to download"
+    exit 0
+else
+    echo "Downloading ${#to_download[@]} files"
+
+    # get files
+    for Experimental_ID in "${to_download[@]}"
+    do
+        
+        outfile="resources/tracks/${genome}/${Experimental_ID}"
+        if [ -e "$outfile" ]
+        then
+            rm "$outfile"
+        fi
+
+        echo "Getting $outfile"
+        ext=$(echo "$outfile" | cut -f2- -d'.')
+        if [ "$ext" = "bw" ]
+        then
+            echo bigwig
+            # BigWig Download URL:
+            wget https://chip-atlas.dbcls.jp/data/"${genome}"/eachData/bw/"${Experimental_ID}" -O "$outfile"
+        fi
+
+        if [ "$ext" = "05.bb" ]
+        then
+            echo "bigbed"
+            # Peak-call (BigBed) Download URL: #(Threshold = 05, 10, or 20)
+            wget https://chip-atlas.dbcls.jp/data/"${genome}"/eachData/bb05/"${Experimental_ID}" -O "$outfile"
+        fi
+    done
     
-    outfile="resources/tracks/${Genome}/${Experimental_ID}"
-    if [ -e "$outfile" ]
-    then
-        rm "$outfile"
-    fi
-
-    echo "Getting $outfile"
-    ext=$(echo "$outfile" | cut -f2- -d'.')
-    if [ "$ext" = "bw" ]
-    then
-        echo bigwig
-        # BigWig Download URL:
-        wget https://chip-atlas.dbcls.jp/data/${Genome}/eachData/bw/${Experimental_ID} -O "$outfile"
-    fi
-
-    if [ "$ext" = "05.bb" ]
-    then
-        echo "bigbed"
-        # Peak-call (BigBed) Download URL: #(Threshold = 05, 10, or 20)
-        wget https://chip-atlas.dbcls.jp/data/${Genome}/eachData/bb05/${Experimental_ID} -O "$outfile"
-    fi
-done
+fi
